Moved glob routing tests into routing_test.rb
